Popup

+++ 3GDBSpeGP/Assets/Editor/WeaponDataTable.cs
@@ -13,12 +13,13 @@ namespace Editor
             window.titleContent = new GUIContent("WeaponDataTable");
             window.Show();
         }
-
         private void OnGUI()
         {
             GUI.backgroundColor = new Color(0.3f, 0.1f, 0f);
             GUILayout.Box("WeaponDataTable", GUILayout.ExpandWidth(true), GUILayout.Height(30));
-            LoadAllAssetsOfType<WeaponData>(out WeaponData[] weapons);
+            
+            LoadAllAssetsOfType(out WeaponData[] weapons);
+
             if (weapons.Length > 0)
             {
                 SerializedObject so = null;
@@ -26,22 +27,26 @@ namespace Editor
                 {
                     so = new SerializedObject(weapon);
                     so.Update();
+
                     EditorGUILayout.BeginHorizontal();
-                    
+
                     GUILayout.Label(weapon.name);
                     GUILayout.FlexibleSpace();
-                    EditorGUILayout.PropertyField(so.FindProperty("FireMode"),GUIContent.none);
-                    EditorGUILayout.PropertyField(so.FindProperty("ShootType"),GUIContent.none);
-                    
-                    if (GUILayout.Button("Create Weapon"))
-                    {
-                        DuplicateWeapon(weapon);
-                    }
+                    EditorGUILayout.PropertyField(so.FindProperty("FireMode"), GUIContent.none);
+                    EditorGUILayout.PropertyField(so.FindProperty("ShootType"), GUIContent.none);
 
-                    if (GUILayout.Button("Delete Weapon"))
-                    {
-                        DeleteWeapon(weapon);
-                    }
+                    EditorGUILayout.PropertyField(so.FindProperty("model"), GUIContent.none);
+                    EditorGUILayout.PropertyField(so.FindProperty("modelSprite"), GUIContent.none);
+
+                    EditorGUILayout.PropertyField(so.FindProperty("damage"), GUIContent.none);
+                    EditorGUILayout.PropertyField(so.FindProperty("accuracy"), GUIContent.none);
+                    EditorGUILayout.PropertyField(so.FindProperty("recoilForce"), GUIContent.none);
+                    EditorGUILayout.PropertyField(so.FindProperty("maxAmmo"), GUIContent.none);
+
+                    if (GUILayout.Button("Create Weapon")) DuplicateWeapon(weapon);
+
+                    if (GUILayout.Button("Delete Weapon")) DeleteWeapon(weapon);
+
                     so.ApplyModifiedProperties();
                     EditorGUILayout.EndHorizontal();
                 }
@@ -51,8 +56,9 @@ namespace Editor
         private void DuplicateWeapon(WeaponData weapon)
         {
             // Duplicate the weapon
-            WeaponData duplicatedWeapon = Instantiate(weapon);
-            AssetDatabase.CreateAsset(duplicatedWeapon, AssetDatabase.GenerateUniqueAssetPath("Assets/Data/Weapons/" + weapon.name + "_Copy.asset"));
+            var duplicatedWeapon = Instantiate(weapon);
+            AssetDatabase.CreateAsset(duplicatedWeapon,
+                AssetDatabase.GenerateUniqueAssetPath("Assets/Data/Weapons/" + weapon.name + "_Copy.asset"));
             AssetDatabase.SaveAssets();
             AssetDatabase.Refresh();
             Debug.Log("Duplicated " + weapon.name);
@@ -60,20 +66,21 @@ namespace Editor
 
         private void DeleteWeapon(WeaponData weapon)
         {
+            CreateWindow<PopupContainer>();
             var path = AssetDatabase.GetAssetPath(weapon);
             AssetDatabase.DeleteAsset(path);
             AssetDatabase.Refresh();
             Debug.Log("Deleted");
         }
-        
+
         private void LoadAllAssetsOfType<T>(out T[] assets) where T : Object
         {
-            string[] guids = AssetDatabase.FindAssets("t:"+typeof(T));
+            var guids = AssetDatabase.FindAssets("t:" + typeof(T));
             assets = new T[guids.Length];
 
-            for (int i = 0; i < guids.Length; i++)
+            for (var i = 0; i < guids.Length; i++)
             {
-                string assetPath = AssetDatabase.GUIDToAssetPath(guids[i]);
+                var assetPath = AssetDatabase.GUIDToAssetPath(guids[i]);
                 assets[i] = AssetDatabase.LoadAssetAtPath<T>(assetPath);
             }
         }
